Move resources used in population/extraction into separate file
