+++ apps/smart-forms-app/src/test/aboriginalFormUtils.tsx
@@ -6,11 +6,12 @@ import {
 } from '@aehrc/smart-forms-renderer';
 
 import aboriginalForm from '../data/resources/Questionnaire/Questionnaire-AboriginalTorresStraitIslanderHealthCheckAssembled-0.4.0.json';
-import type { Questionnaire } from 'fhir/r4';
+import type { Questionnaire, QuestionnaireResponse } from 'fhir/r4';
 import { QueryClientProvider } from '@tanstack/react-query';
 import type { Patient } from 'fhir/r4';
 import { populateQuestionnaire } from '@aehrc/sdc-populate';
 import { useEffect, useState } from 'react';
+import { extractResultIsOperationOutcome, inAppExtract } from '@aehrc/sdc-template-extract';
 const terminologyServerUrl = 'https://r4.ontoserver.csiro.au/fhir';
 
 export type RequestDefinition = {
@@ -124,4 +125,22 @@ function buildFetchResourceCallback(requestDefinitions: RequestDefinition[]) {
 
     return Promise.resolve({});
   };
+}
+
+export async function runExtract(
+  questionnaireResponse: QuestionnaireResponse,
+  questionnaire: Questionnaire
+) {
+  const inAppExtractOutput = await inAppExtract(questionnaireResponse, questionnaire, null);
+
+  if (!inAppExtractOutput.extractSuccess) {
+    throw new Error('Extract failed');
+  }
+
+  const extractResult = inAppExtractOutput.extractResult;
+  if (extractResultIsOperationOutcome(extractResult)) {
+    throw new Error('Extract failed');
+  }
+
+  return extractResult.extractedBundle;
 }
